Put the virtual CPU trait in a separate module
